output: bind new modes in the order we add them
Currently, when a new mode is bound to an output, it is inserted at the
head of the list, i.e. reversed from the order they are present in the
drmModeConnector modes array. But it's better to keep the order, because
it seems the best (high-res) modes come first; this way the lazy user,
who just passes NULL to kmscon_output_activate, would get a nice looking
default mode.