font: draw backgrounds
The pango backend now correctly draws colored or uncolored backgrounds.
This is needed to support drawing cursors.

We currently do not optimize this for performance. In fact, this reduces
performance a lot and we could avoid drawing backgrounds if we cleared the
background to the same color before. However, this also means we currently
could move drawing the background into the console backend again. The
transparency feature is awful, anyway and I don't see any reason
supporting it.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>