uterm: vt: silence gcc warning
We initialize the variable to "true" to silence gcc warnings that we might
get an uninitialized variable (which we really cannot get here).

Anyway, in the future, if the number of types may increase, the correct
way to deal with unknown types is returning "true" so do this by default
now.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>