ui: arrange video objects in a list
If multiple video objects are added to a UI object, we need to manage them
in a list to allow each of them to be added to a running terminal.

This only changes the internal structure but does still create the video
object when creating the UI object. This is, because the terminal
subsystem depends on uterm-video and this needs to be resolved, first.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>