fakevt: add fakevt helper for VT-less systems
The fakevt helper binary can be used on VT-less systems or seats to make
kmscon activate the fake VT and deactivate it. This way, you can control
when kmscon acquires video devices and when it releases them.

This is a global setting that affects all seats where kmscon is running
except seat0 if it uses real VTs.

This should only be used for debugging. This is really no intended for use
in production.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>